Moved commands to their own submodules with their newly added unimplemented parsers
